data-engineering/whale_api/requirements.txt
# python library dependencies
types-requests
python-dotenv
sqlalchemy
requests
psycopg2-binary
pandas
mypy